changes to limit analysis project disk usage report to reflect just the usage of that project and to display additional precision in the TB display value

+++ app/controllers/analysis_projects_controller.rb
@@ -20,7 +20,7 @@ class AnalysisProjectsController < ApplicationController
                                                      .order("updated_at DESC")
                                                      .limit(25)
     @analysis_project_disk_usage = Rails.cache.fetch("analysis project disk usage query" + params[:id]) do
-     AnalysisProjectDiskUsageQuery.new.execute(params[:id]).to_a[0]['usage'].to_i/(1024*1024*1024)
+     AnalysisProjectDiskUsageQuery.new.execute(params[:id]).to_a[0]['usage'].to_f/(1024*1024*1024)
     end
   end
